- rename and moves re-usable interest selection components to compose/components
