+++ app/src/main/java/org/wikipedia/compose/components/SearchBarCard.kt
@@ -1,0 +1,57 @@
+package org.wikipedia.compose.components
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import org.wikipedia.R
+import org.wikipedia.compose.theme.WikipediaTheme
+
+@Composable
+fun SearchBarCard(
+    modifier: Modifier = Modifier,
+    onSearchClick: () -> Unit
+) {
+    Row(
+        modifier = modifier
+            .fillMaxWidth()
+            .height(56.dp)
+            .clip(RoundedCornerShape(28.dp))
+            .background(
+                color = WikipediaTheme.colors.backgroundColor,
+                shape = RoundedCornerShape(24.dp)
+            )
+            .clickable(onClick = onSearchClick),
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        Spacer(modifier = Modifier.width(16.dp))
+        Icon(
+            painter = painterResource(R.drawable.outline_search_24),
+            contentDescription = stringResource(R.string.search_hint),
+            tint = WikipediaTheme.colors.secondaryColor,
+            modifier = Modifier.size(24.dp)
+        )
+        Text(
+            modifier = Modifier.padding(start = 16.dp, end = 16.dp),
+            text = stringResource(R.string.recommended_reading_list_interest_pick_search_hint),
+            style = MaterialTheme.typography.bodyLarge,
+            color = WikipediaTheme.colors.primaryColor
+        )
+    }
+}
